Scene_Text_Detection_and_Recognition/requirements.txt
tensorflow==1.14
numpy==1.19.5
streamlit==0.80.0
matplotlib==3.2.2 

seaborn==0.11.1 

scikit-learn==0.22.2.post1 
Keras==2.4.3         
Keras-Preprocessing==1.1.2 

pytesseract== 0.3.7
pyspellchecker==0.6.2  

opencv-python-headless
Shapely





